backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
psycopg2-binary==2.9.10
Pillow==11.1.0
django-environ==0.11.2
drf-spectacular==0.28.0
